pyproject.toml
[project]
name = "paper-trader"
version = "0.1.0"
description = "Paper trading bot with AI-powered decisions"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "aiosqlite>=0.21",
    "yfinance>=0.2",
    "feedparser>=6.0",
    "anthropic>=0.52",
    "apscheduler>=3.11,<4",
    "jinja2>=3.1",
    "pydantic-settings>=2.7",
    "httpx>=0.28",
    "python-multipart>=0.0.18",
]

[project.scripts]
paper-trader = "paper_trader.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/paper_trader"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.25",
    "pytest-httpx>=0.35",
    "respx>=0.22",
]
